Created Methodology class. Removed unneeded modules
