+++ tests/test_safelist.py
@@ -9,7 +9,7 @@ of this license document, but changing it is not allowed.
 
 import pytest
 
-from core.safelist import Safelist
+from main.safelist import Safelist
 from slips_aip_constants.defaults import DefaultSafelists
 
 CURRENT_IP_SAFELIST_SIZE = 107
